Refactor ActivityItem and fix theme object refs
