+++ src/fora-components/ActivityItem/Submission.tsx
@@ -1,0 +1,27 @@
+/** @jsx jsx */
+import { jsx } from "theme-ui";
+import React from "react";
+import { Text, Link } from "@theme-ui/components";
+import Description from "./Description";
+
+export interface ISubmissionProps {
+  activityType: "submission";
+  authorName: string | undefined;
+  bountyTitle: string;
+  href?: string;
+}
+
+const Submission: React.FC<ISubmissionProps> = ({
+  authorName,
+  bountyTitle,
+  href
+}) => (
+  <Description authorName={authorName}>
+    <Text mx="1">{`submitted to`}</Text>
+    <Link href={href} sx={{ fontWeight: "medium" }}>
+      {bountyTitle}
+    </Link>
+  </Description>
+);
+
+export default Submission;
